+++ vite.config.ts
@@ -1,8 +1,8 @@
 // Plugins
 import vue from '@vitejs/plugin-vue';
-import vuetify from 'vite-plugin-vuetify';
 import viteCompression from 'vite-plugin-compression';
 import { VitePWA } from 'vite-plugin-pwa';
+import vuetify, { transformAssetUrls } from 'vite-plugin-vuetify';
 import type { VitePWAOptions } from 'vite-plugin-pwa';
 import AutoImport from 'unplugin-auto-import/vite';
 
@@ -51,7 +51,9 @@ const pwaOptions: Partial<VitePWAOptions> = {
 // https://vitejs.dev/config/
 export default defineConfig({
 	plugins: [
-		vue(),
+		vue({
+			template: { transformAssetUrls }
+		}),
 		// https://github.com/vuetifyjs/vuetify-loader/tree/next/packages/vite-plugin
 		vuetify({
 			autoImport: true,