+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# <a href='https://github.com/mrjackwills/leafcast_vue/releases/tag/v0.9.1'>v0.9.1</a>
+### 2022-11-05
+
+### Chores
++ dependencies updated, [54ca70fe](https://github.com/mrjackwills/leafcast_vue/commit/54ca70fe1df00326f8e8671f893244152d119852),
+
+### Fixes
++ app.vue footer height overlay fix, [2a752e95](https://github.com/mrjackwills/leafcast_vue/commit/2a752e95d5e44669141ae4bad01f71cbc9c2263d),
++ img loading via vite template, [56779218](https://github.com/mrjackwills/leafcast_vue/commit/56779218e97bd12522f497a7a66ce0ab5e445d11),
++ update button disabled on pi offline, [29049364](https://github.com/mrjackwills/leafcast_vue/commit/29049364953d6fa444035bb505e2e90ae34eaa78),
++ computed piOnline .value, [48f996e6](https://github.com/mrjackwills/leafcast_vue/commit/48f996e6e119ac0880f1c58253ba92029256143c),, [bac92204](https://github.com/mrjackwills/leafcast_vue/commit/bac92204739c1d260b4c3db22f806f01ab869429),
++ start update countdown even if offline, [3851ad63](https://github.com/mrjackwills/leafcast_vue/commit/3851ad632a3f71c6160b6ab8664fb1788711ae3c),
+
 # <a href='https://github.com/mrjackwills/leafcast_vue/releases/tag/v0.9.0'>v0.9.0</a>
 ### 2022-11-03
 
@@ -22,7 +35,7 @@
 ### 2022-05-30
 
 ### Chores
-+ npm update, [c9e67167](https://github.com/mrjackwills/leafcast_vue/commit/c9e67167697c764460fcbc9ac638adde7dc4b679),, [e0e53dae](https://github.com/mrjackwills/leafcast_vue/commit/e0e53daeb364be09337c265594704b0cb0911b40),
++ npm update, [c9e67167](https://github.com/mrjackwills/leafcast_vue/commit/c9e67167697c764460fcbc9ac638adde7dc4b679), [e0e53dae](https://github.com/mrjackwills/leafcast_vue/commit/e0e53daeb364be09337c265594704b0cb0911b40),
 
 ### Fixes
 + axios 429 response error type work-around, [4d417246](https://github.com/mrjackwills/leafcast_vue/commit/4d417246180c99f9fe25c6f829c83ff9c1076c81),